feat(bootstrap): platform-split self-install with Unix support
Split the Windows-only bootstrap command into platform files. The current
Windows behavior moves byte-identically into bootstrap_windows.go; a new
bootstrap_unix.go implements Linux/macOS:

- install to ${XDG_DATA_HOME:-$HOME/.local/share}/endstate/bin, copy binary to
  lib/endstate and chmod 0755 (the copy is non-executable by default)
- symlink $HOME/.local/bin/endstate -> the installed binary, re-pointed
  idempotently (existing target removed first)
- never edit PATH or shell rc files; addedToPath is always false; a one-line
  PATH hint is printed to stderr when $HOME/.local/bin is off PATH (JSON shape
  unchanged)
- self-copy guard: skip the copy when the running binary already resolves to the
  install target, so re-bootstrapping via the installed symlink never truncates it

The BootstrapData payload (installPath/shimPath/addedToPath) is preserved on every
platform; on Unix shimPath is the symlink path.
