+++ go-engine/internal/commands/bootstrap.go
@@ -4,14 +4,8 @@
 package commands
 
 import (
-	"fmt"
 	"io"
 	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
-
-	"github.com/Artexis10/endstate/go-engine/internal/envelope"
 )
 
 // BootstrapFlags holds the parsed CLI flags for the bootstrap command.
@@ -25,95 +19,20 @@ type BootstrapFlags struct {
 
 // BootstrapData is the data payload returned by the bootstrap command. It
 // reports the paths created and whether the user PATH was modified.
+//
+// On Windows, ShimPath is the path of the generated endstate.cmd shim. On
+// Unix (linux/darwin), ShimPath is the path of the symlink that points at the
+// installed binary, and AddedToPath is always false (the engine never edits
+// shell rc files).
 type BootstrapData struct {
 	InstallPath string `json:"installPath"`
 	ShimPath    string `json:"shimPath"`
 	AddedToPath bool   `json:"addedToPath"`
 }
 
-// shimContent is the endstate.cmd shim that delegates to the Go binary.
-const shimContent = `@echo off
-set "ENDSTATE_ENTRYPOINT=%~f0"
-"%~dp0lib\endstate.exe" %*
-`
-
-// RunBootstrap installs the running binary to the user-local Endstate
-// directory and creates a .cmd shim on the user PATH.
-//
-// Steps:
-//  1. Determine install dir: %LOCALAPPDATA%\Endstate\bin\
-//  2. Create install dir and lib\ subdirectory.
-//  3. Copy running binary to lib\endstate.exe.
-//  4. Write endstate.cmd shim to install dir.
-//  5. Add install dir to user PATH if not already present.
-func RunBootstrap(flags BootstrapFlags) (interface{}, *envelope.Error) {
-	installDir := os.ExpandEnv("${LOCALAPPDATA}\\Endstate\\bin")
-	libDir := filepath.Join(installDir, "lib")
-
-	// Create directories.
-	if err := os.MkdirAll(libDir, 0755); err != nil {
-		return nil, envelope.NewError(
-			envelope.ErrInternalError,
-			fmt.Sprintf("Failed to create install directory: %s", err.Error()),
-		)
-	}
-
-	// Get the path of the currently running binary.
-	exePath, err := os.Executable()
-	if err != nil {
-		return nil, envelope.NewError(
-			envelope.ErrInternalError,
-			fmt.Sprintf("Failed to determine executable path: %s", err.Error()),
-		)
-	}
-
-	// Resolve symlinks to get the real path.
-	exePath, err = filepath.EvalSymlinks(exePath)
-	if err != nil {
-		return nil, envelope.NewError(
-			envelope.ErrInternalError,
-			fmt.Sprintf("Failed to resolve executable path: %s", err.Error()),
-		)
-	}
-
-	// Copy binary to lib\endstate.exe.
-	destBinary := filepath.Join(libDir, "endstate.exe")
-	if err := copyFile(exePath, destBinary); err != nil {
-		return nil, envelope.NewError(
-			envelope.ErrInternalError,
-			fmt.Sprintf("Failed to copy binary: %s", err.Error()),
-		)
-	}
-
-	// Write the .cmd shim.
-	shimPath := filepath.Join(installDir, "endstate.cmd")
-	if err := os.WriteFile(shimPath, []byte(shimContent), 0755); err != nil {
-		return nil, envelope.NewError(
-			envelope.ErrInternalError,
-			fmt.Sprintf("Failed to write shim: %s", err.Error()),
-		)
-	}
-
-	// Check if installDir is already in the user PATH and add it if not.
-	addedToPath := false
-	if !isInUserPath(installDir) {
-		if err := addToUserPath(installDir); err != nil {
-			// PATH modification failure is not fatal; report it but succeed.
-			// The user can add it manually.
-			_ = err
-		} else {
-			addedToPath = true
-		}
-	}
-
-	return &BootstrapData{
-		InstallPath: installDir,
-		ShimPath:    shimPath,
-		AddedToPath: addedToPath,
-	}, nil
-}
-
-// copyFile copies src to dst, overwriting dst if it exists.
+// copyFile copies src to dst, overwriting dst if it exists. The destination is
+// created with the default os.Create mode (0666 before umask); callers that
+// need the destination to be executable adjust its mode after the copy.
 func copyFile(src, dst string) error {
 	in, err := os.Open(src)
 	if err != nil {
@@ -132,29 +51,4 @@ func copyFile(src, dst string) error {
 	}
 
 	return out.Close()
-}
-
-// isInUserPath checks whether dir is present in the current PATH environment
-// variable (case-insensitive on Windows).
-func isInUserPath(dir string) bool {
-	pathEnv := os.Getenv("PATH")
-	normalised := strings.ToLower(filepath.Clean(dir))
-
-	for _, entry := range filepath.SplitList(pathEnv) {
-		if strings.ToLower(filepath.Clean(entry)) == normalised {
-			return true
-		}
-	}
-	return false
-}
-
-// addToUserPath appends dir to the user PATH via setx. This persists the
-// change to the registry (HKCU\Environment\Path) and takes effect in new
-// shell sessions.
-func addToUserPath(dir string) error {
-	currentPath := os.Getenv("PATH")
-	newPath := currentPath + ";" + dir
-
-	cmd := exec.Command("setx", "PATH", newPath)
-	return cmd.Run()
 }
